vm-config/requirements.txt
streamlit
openai
azure-identity
azure-keyvault-secrets